Fixed occasional gesture lag
When the full screen view is opened, occasionally it will be unresponsive to gestures that are submitted quickly after opening (e.g. double tap or pinch to zoom). This issue was caused by accessory view animation not allowing UserInteraction. Saw debug log messages like this every time the issue was reproduced:

    Gesture: Failed to receive system gesture state notification before next touch
    Failed to receive system gesture state notification before next touch